requirements.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.3.4
billiard==3.3.0.23
boto==2.42.0
cairocffi==0.7.2
CairoSVG==1.0.22
celery==3.1.18
cffi==1.8.3
cssselect==1.0.0
dj-database-url==0.4.1
Django==1.8.5
django-embed-video==1.1.0
django-localflavor==1.1
django-parler==1.5.1
django-paypal==0.2.5
django-rosetta==0.7.6
django-storages==1.5.1
flower==0.9.1
gunicorn==19.6.0
html5lib==0.999999999
kombu==3.0.36
lxml==3.6.4
microsofttranslator==0.5
Pillow==2.9.0
polib==1.0.7
psycopg2==2.6.2
pycparser==2.17
Pyphen==0.9.4
pytz==2016.7
redis==2.10.3
requests==2.11.1
six==1.10.0
South==1.0.2
tinycss==0.3
tornado==4.2
WeasyPrint==0.24
webencodings==0.5
whitenoise==3.2.1
